backend/pyproject.toml
[tool.poetry]
name = "ticketing-backend"
version = "0.1.0"
description = "FastAPI backend for EventTicketing dApp"
authors = ["ANTIGRAVITY"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
algokit-utils = "^4.0.0"
algosdk = "^3.0.0"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"
black = "*"
ruff = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
